app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
structlog==24.2.0
google-cloud-secret-manager==2.20.0
google-cloud-monitoring==2.22.0
protobuf==4.25.3
httpx==0.27.0
